[RunnableConfigObject] Use shell execution
Allow the user to make use of shell features in commands.  The user will expect
to be able to construct command pipelines, use redirection, etc. and attempting
to parse shell features and translate them to library calls is not only a
gigantic project, it arguably doesn't add any security protection over using
shell execution, since we would just translate the command to do exactly the
actions the user specifies anyway.