[ConfigObject] Allow user-settable config values
Make ConfigObject more versatile by freeing it from the ConfigObjectItem method
of tracking git config entries.  Allow the user to set any attribute and have it
saved to the git config for later retrieval.

Names that begin with underscores are assumed to be private and are not recorded
in the git config.