[Git] Don't duplicate config entries
When initializing a ConfigSection, don't iteratively add key, value pairs.
Since we're reading values from the underlying config file, we don't need to
add_items to the config again.  Instead, generate a dictionary for each section
and initialize the sections in one go, without updating the underlying config.

ConfigSection is really just a cache of the underlying config so initializing it
should not change the underlying config.