[main] Translate illegal config keys
Since the project name is used as a config key, it must adhere to git's rules
about config keys.  If the project happens to contain . (which would create a
config section) or _ (which is illegal), translate them to -.

If the resulting string starts with -, prepend a string to make it legal.
Technically git allows a section name to begin with - but pygit2 does not.