[substitutable] Allow more advanced substitution
Use eval to allow the user to format a string arbitrarily.  Try to limit eval
damage by requiring that the eval only work with keys in the constructed
substitution dictonary.

We have to use eval because there is no way for an f-string to evaluate itself
as an f-string.  Regular string.format doesn't allow string methods to be called
(i.e. string.format('{var.strip()}') doesn't work).  f-strings allow that but to
use f-strings we need a string literal and what we have is a string read from a
config file.  The solution is to first substitute that string into an f-string
and then eval that as another f-string.

It is possible, maybe even probable, that the string read from the config could
do something nefarious.  Using the keyword dictionary as the global namespace
disallows obvious things like '{os.rmdir($HOME)}' since 'os' won't be in the
dictionary but a careful user could probably set something like that up in the
config by setting 'branch' to 'os.rmdir($HOME)' so that referencing '{branch}'
would do it.  For now we trust that the user won't do that and won't blindly
trust third-party scripts that set up a git-project config.