[substitutable] Allow escape of braces
The python f-string brace escape sequence {{ and }} doesn't work for
substitutables since we keep substituting until there is no change.  The {{ }}
gets translated to { } which can itself be a valid substitution.

Allow the user to use the somewhat odd {{} and {}} to escape braces.  These get
replaced temporarily by an equally unlikely string and then are restored to
single braces at the end.  If a user truly wants that sequence to appear
verbatim, they can escape each brace.